requirements.txt
streamlit>=1.35
xgboost>=2.0
shap>=0.45
scikit-learn>=1.4
pandas>=2.0
numpy>=1.26
joblib>=1.3
matplotlib>=3.8
openai>=1.30
python-dotenv>=1.0
